reference_streamlit/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==24.1.0
    # via gradio
altair==6.0.0
    # via streamlit
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   gradio
    #   httpx
    #   starlette
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
audioop-lts==0.2.2
    # via gradio
bcrypt==5.0.0
    # via -r requirements.in
blinker==1.9.0
    # via streamlit
brotli==1.2.0
    # via gradio
cachetools==6.2.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   pdfminer-six
    #   requests
click==8.3.1
    # via
    #   nltk
    #   streamlit
    #   typer
    #   typer-slim
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==46.0.3
    # via pdfminer-six
defusedxml==0.7.1
    # via fpdf2
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.124.0
    # via gradio
ffmpy==1.0.0
    # via gradio
filelock==3.20.0
    # via
    #   huggingface-hub
    #   transformers
fonttools==4.61.0
    # via fpdf2
fpdf2==2.8.5
    # via -r requirements.in
fsspec==2025.12.0
    # via
    #   gradio-client
    #   huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core[grpc]==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.8.5
    # via -r requirements.in
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
gradio==6.0.2
    # via -r requirements.in
gradio-client==2.0.1
    # via gradio
groovy==0.1.2
    # via gradio
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   safehttpx
huggingface-hub==0.36.0
    # via
    #   gradio
    #   gradio-client
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   gradio
    #   pydeck
joblib==1.5.2
    # via nltk
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
lxml==6.0.2
    # via python-docx
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==2.13.0
    # via
    #   altair
    #   plotly
nltk==3.9.2
    # via -r requirements.in
numpy==2.2.6
    # via
    #   gradio
    #   opencv-python
    #   pandas
    #   pydeck
    #   streamlit
    #   transformers
opencv-python==4.12.0.88
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
orjson==3.11.5
    # via gradio
packaging==25.0
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   plotly
    #   pytesseract
    #   streamlit
    #   transformers
pandas==2.3.3
    # via
    #   -r requirements.in
    #   gradio
    #   streamlit
pdf2image==1.17.0
    # via -r requirements.in
pdfminer-six==20251107
    # via
    #   -r requirements.in
    #   pdfplumber
pdfplumber==0.11.8
    # via -r requirements.in
pillow==12.0.0
    # via
    #   -r requirements.in
    #   fpdf2
    #   gradio
    #   pdf2image
    #   pdfplumber
    #   pytesseract
    #   streamlit
plotly==6.5.0
    # via -r requirements.in
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==22.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   fastapi
    #   google-generativeai
    #   gradio
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pydub==0.25.1
    # via gradio
pygments==2.19.2
    # via rich
pymysql==1.1.2
    # via -r requirements.in
pyparsing==3.2.5
    # via httplib2
pypdfium2==5.1.0
    # via pdfplumber
pytesseract==0.3.13
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.2.0
    # via -r requirements.in
python-multipart==0.0.20
    # via gradio
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   gradio
    #   huggingface-hub
    #   transformers
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.11.3
    # via
    #   nltk
    #   transformers
requests==2.32.5
    # via
    #   google-api-core
    #   huggingface-hub
    #   streamlit
    #   transformers
rich==14.2.0
    # via typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
safehttpx==0.1.7
    # via gradio
safetensors==0.7.0
    # via transformers
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
starlette==0.50.0
    # via
    #   fastapi
    #   gradio
streamlit==1.52.1
    # via
    #   -r requirements.in
    #   streamlit-tags
streamlit-tags==1.2.8
    # via -r requirements.in
tenacity==9.1.2
    # via streamlit
tokenizers==0.22.1
    # via
    #   -r requirements.in
    #   transformers
toml==0.10.2
    # via streamlit
tomlkit==0.13.3
    # via gradio
tornado==6.5.2
    # via streamlit
tqdm==4.67.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   nltk
    #   transformers
transformers==4.57.3
    # via -r requirements.in
typer==0.20.0
    # via gradio
typing-extensions==4.15.0
    # via
    #   altair
    #   fastapi
    #   google-generativeai
    #   gradio
    #   gradio-client
    #   grpcio
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   streamlit
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.1
    # via requests
uvicorn==0.38.0
    # via gradio
watchdog==6.0.0
    # via streamlit
yt-dlp==2025.12.8
    # via -r requirements.in
